flask-app/requirements.txt
aws-xray-sdk
Babel
boto3
botocore
certifi
chardet
click
docutils
dominate
Flask
Flask-Babel
Flask-Bootstrap4
Flask-WTF
future
idna
itsdangerous
Jinja2
jmespath
jsonpickle
MarkupSafe
python-dateutil
python-dotenv
pytz
requests
s3transfer
six
urllib3
visitor
Werkzeug
wrapt
WTForms
